Menghapus class dengan tugas terlalu sedikit
